refactor(squad_selection): reorganise modules
