+++ fpl_predictor/squad_selection/__init__.py
@@ -1,0 +1,3 @@
+from .squad_selection import select_squad
+
+__all__ = ["select_squad"]
